Crud medicine and add table category and supplier and add feature cart

+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\CartController;
 use App\Http\Controllers\MedicineController;
 use App\Http\Controllers\ProfileController;
 use App\Http\Middleware\CheckRole;
@@ -8,6 +9,7 @@ use App\Models\Employee;
 use App\Models\Medicine;
 use App\Models\Restock;
 use App\Models\Sale;
+use App\Models\Supplier;
 use App\Models\User;
 use Illuminate\Support\Facades\Route;
 
@@ -24,7 +26,10 @@ Route::get('/cashier', function () {
 })->middleware(['auth', 'verified', CheckRole::class . ':admin,pharmacist,cashier,inventory_manager'])->name('cashier');
 
 Route::get('/medicines', function () {
-    return view('medicines', ['medicines' => Medicine::latest()->get()]);
+    // Mengambil data medicines dengan eager loading untuk kategori dan supplier
+    $medicines = Medicine::with(['category', 'supplier'])->latest()->get();
+
+    return view('medicines', ['medicines' => $medicines]);
 })->middleware(['auth', 'verified', CheckRole::class . ':admin,pharmacist,inventory_manager'])->name('medicines');
 
 Route::get('/medicines/{id}', function ($id) {
@@ -48,8 +53,8 @@ Route::get('/customers', function () {
 })->middleware(['auth', 'verified', CheckRole::class . ':admin,pharmacist,cashier'])->name('customers');
 
 Route::get('/suppliers', function () {
-    $restocks = Restock::with('medicine')->get(); // eager loading
-    return view('suppliers', ['restocks' => $restocks]);
+    $suppliers = Supplier::with('medicines')->get(); // eager loading
+    return view('suppliers', ['suppliers' => $suppliers]);
 })->middleware(['auth', 'verified', CheckRole::class . ':admin,inventory_manager'])->name('suppliers');
 
 Route::get('/employees', function () {
@@ -60,9 +65,13 @@ Route::get('/users', function () {
     return view('users', ['users' => User::all()]);
 })->middleware(['auth', 'verified', CheckRole::class . ':admin'])->name('users');
 
-Route::post('/medicines', [MedicineController::class, 'store'])->name('medicine.add');
-Route::delete('/medicines/{id}', [MedicineController::class, 'destroy'])->name('medicine.delete');
-Route::put('/medicines/{id}', [MedicineController::class, 'update'])->name('medicine.update');
+// Medicine CRUD
+Route::post('/medicines', [MedicineController::class, 'store'])->name('medicine.add')->middleware(['auth', 'verified', CheckRole::class . ':admin,pharmacist,inventory_manager']);  
+Route::delete('/medicines/{id}', [MedicineController::class, 'destroy'])->name('medicine.delete')->middleware(['auth', 'verified', CheckRole::class . ':admin,pharmacist,inventory_manager']);
+Route::put('/medicines/{id}', [MedicineController::class, 'update'])->name('medicine.update')->middleware(['auth', 'verified', CheckRole::class . ':admin,pharmacist,inventory_manager']);
+Route::post('/medicines/search', [MedicineController::class, 'search'])->name('medicine.search')->middleware(['auth', 'verified', CheckRole::class . ':admin,pharmacist,inventory_manager']);
+
+// sales CRUD
 
 Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
